Reorganizing files into folders, no functional changes.
